feat: PostInspector 추가 (#96)
