--- src/main/java/com/wegotoo/config/EmbeddedRedisConfig.java
@@ -1,78 +1,0 @@
-package com.wegotoo.config;
-
-import static io.micrometer.common.util.StringUtils.isEmpty;
-
-import jakarta.annotation.PostConstruct;
-import jakarta.annotation.PreDestroy;
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.data.redis.RedisProperties;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Profile;
-import redis.embedded.RedisServer;
-
-@Configuration
-@RequiredArgsConstructor
-@Slf4j
-@Profile("local")
-public class EmbeddedRedisConfig {
-
-    private final RedisProperties redisProperties;
-
-    private RedisServer redisServer;
-
-    @PostConstruct
-    private void start() throws IOException {
-        int redisPort = isPortRunning() ? findAvailablePort() : redisProperties.getPort();
-        redisProperties.setPort(redisPort);
-        redisServer = new RedisServer(redisPort);
-        redisServer.start();
-    }
-
-    @PreDestroy
-    private void stop() throws IOException {
-        if (redisServer != null) {
-            redisServer.stop();
-        }
-    }
-
-    private boolean isPortRunning() throws IOException {
-        return isRunning(executeGrepProcessCommand(redisProperties.getPort()));
-    }
-
-    private int findAvailablePort() throws IOException {
-        for (int port = 10000; port <= 65535; port++) {
-            Process process = executeGrepProcessCommand(port);
-
-            if (!isRunning(process)) {
-                return port;
-            }
-        }
-
-        throw new IllegalArgumentException("Not Found Available Port: 10000 ~ 65535");
-    }
-
-    private Process executeGrepProcessCommand(int port) throws IOException {
-        String command = String.format("netstat -nat | grep LISTEN|grep %d", port);
-        String[] shell = {"/bin/sh", "-c", command};
-        return Runtime.getRuntime().exec(shell);
-    }
-
-    private boolean isRunning(Process process) {
-        String line;
-        StringBuilder pidInfo = new StringBuilder();
-
-        try (BufferedReader input = new BufferedReader(new InputStreamReader(process.getInputStream()))) {
-            while ((line = input.readLine()) != null) {
-                pidInfo.append(line);
-            }
-        } catch (Exception e) {
-        }
-
-        return !isEmpty(pidInfo.toString());
-    }
-
-}
